refactor/gh-40
- Move helper function 'validate_search_params' in 'listings'
  app from views.py to functions.py.
